iio: frequency: dds: Unset ad3552r issue_sync_en flag
The DDS will not write the SYNC code to REG_CNTRL_1 if DMA
synchronization is needed or if the issue_sync_en flag is set.
We might actually need to keep this flag set because ad3552r HDL
projects provide external signal to sync the DAC, but let's experiment
with it to see what happens.