[WIP]: adaq4224 as inkernel consumer
Provide support to ADAQ4224 as a wrapper driver that consumes AD4030 and
MAX31827 channels.
The possible advantage of this implementation is flexibility
in case new products come with different ADCs or temperature sensors
integrated we would simply change the configuration in device tree
and have all the channel grouped under a same driver again.
It is not clear though how we will handle buffers and there might be
problems if concurrent access to adaq4224 and ad4030 happen.

ADAQ_ADC: rename adaq4224.c -> adaq_adc

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>