iio: dac: ad3552r: Handle read transfer entirely on reg_read
Having a single transfer function for both writing and reading to/from
the device forces us to do additional checks for each case, leading to
intricate code.
The only place the transfer function was called to read from the device
was at ad3552r_read_reg(). Rework the register read call to fully handle
device read operations so to have separate functions for reads and writes.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt1@gmail.com>