iio: frequency: cf_axi_dds: Allow linked devices to register clk notifier
Create a device tree property called adi,shared-clock and use it to
share the linked converter frontend device reference clock with the dds
core driver. This is important if the converter device is required to
allow sampling frequency configuration and the clock reference is shared
between the dds core and converter device.
This allows the dds core to adjust the frequency of the phase accumulator
when a converter sampling rate change leads to a clock frequency change.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>