iio: adc: ad400x: Remove unnecessary/wrong delay
word_delay is not the proper delay for AD400x transfers.
As for the documentation in spi.h, word_delay is a inter word delay
introduced after each word size transmission.
However, the timing specification intended to be met here is tQUIET2
(60 ns) which is the time between the last SCLK FALLING EDGE and CNV
(CS) rising edge. That is more similar to the delay parameter which will
be provided on the following commit.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>