ad4000: This reads samples correctly, reg read/writes work as by vivado
This reads samples correctly, reg read/writes are coherent with
signals observed on vivado with ILA debug cores and seem to be correct
according to datasheet.
Actual register read data, however, didn't seem correct since it looked like
it was floating. May the eval board was in an unstable state.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>