iio: adc: ad_sigma_delta: add disable_one callback
Sigma delta ADCs with a sequencer need to disable the previously enabled
channel when reading using ad_sigma_delta_single_conversion(). This was
done manually in drivers for devices with sequencers.

This patch implements handling of single channel disabling after a
single conversion.

Reviewed-by: Nuno Sa <nuno.sa@analog.com>
Signed-off-by: Dumitru Ceclan <dumitru.ceclan@analog.com>
Reviewed-by: David Lechner <dlechner@baylibre.com>
Link: https://patch.msgid.link/20240607-ad4111-v7-3-97e3855900a0@analog.com
Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>

iio: adc: ad_sigma_delta: fix disable_one callback

The ADC ad7192 is a sigma delta ADC with a sequencer that does not
require a disable_one callback as all enable channel bits are within
the same register.

Remove the requirement of the disable_one callback for sigma delta ADCs
with a sequencer.

This patch could be squashed with the commit that it fixes from patch
series: Add support for AD411x

Fixes: a25a0aab2187 ("iio: adc: ad_sigma_delta: add disable_one callback")
Signed-off-by: Dumitru Ceclan <dumitru.ceclan@analog.com>
Link: https://lore.kernel.org/r/20240621120539.31351-1-dumitru.ceclan@analog.com
Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>