iio: adc: ad4170: Fix ADC scale update
The scale stored in scale_tbl is the same one that is displaied to user
space through _scale_available channel attribute. When updating the ADC
channel scale, the scale fractional part should match a value in scale_tbl.
Previously, the scale franctional part was being devided by 1000 and that
was preventing the scale update.
Do not divide the input scale fractional part so it can match values in
scale_tbl and the scale can be updated correctly.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>