iio: adc: ad4170: Review transfer buffers - split message into 2 SPI transfers
Buffered reads all contained 45 as MSB byte which is the second byte of the
synchronization pattern which is 0x2645.
The instruction phase is 16-bit and the data phase would be 24-bit because
DATA_24B (0x1E) is 24-bit register. That leads to 16 + 24 = 48 bit transfer
which is odd to handle as a single SPI transfer.
Split the SPI message into two SPI transfers (one for the instruction
phase, other for the data phase).
Experimental change to be tested.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>