iio: adc: ad4170: Fix buffer configuration for bipolar channels
The channel template was being initialized with buffer configured for
unipolar (straight binary) data. Even though channel parse and
initialization was assigning configuration for signed values for bipolar
channels, the final channel configuration was not changing due to the const
modifier of the channel template. Do not initialize the channel template
with unsigned buffer element configuration. Set the buffer sign property on
channel initialization only.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>