dt-bindings: iio: adc: adi,ad4170: Drop adi,power-down-switch-gpio/pin
It's not clear how to document adi,power-down-switch-gpio/pin in device tree
in a way that would be guarantee to age well as other/future designs might
have different ways to setup bridge circuits power-down-switches. Also,
I'm not aware of anybody requesting support for this specific feature so we
can probably extend the bindings if/when somebody requests that.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>