iio: adc: ad4170: Fix the filling of IIO buffer
The documentation (Documentation/iio/iio_devbuf.rst) is correct and the IIO
buffer data is expected to be aligned according to the size and number of
enabled IIO channels. Thus, when lower index channels are disabled,
channels with higher indexes ought to be stored closer to the beginning of
the IIO buffer. Fill the IIO buffer correctly by copying data from it's
start to it's end in incremental order of enabled channels instead of
storing channel data at buffer the element that equals the channel index.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>