iio: buffer: industrialio-buffer-dmaengine: [Experimental] Discard half of data buffer
Hack to discard ADAQ4216 sencond channel data.
The ad4630_fmc HDL project was designed for ADCs with two channels
and always stream two data channels to DMA (even when the ADC has
only one physical channel). Though, if the ADC has only one physical
channel, the data that would be from the second ADC channel comes in
as noise and has to be discarded. Because of that, only half of data
comming from DMA is used for single channel ADCs such as ADAQ4216.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>