iio: buffer: Add industrialio-buffer-dmaengine-filtered
This is a hack to discard ADAQ4216 second channel data.

The ad4630_fmc HDL project was designed for ADCs with two channels and
always streams two data channels to DMA (even when the ADC has only one
physical channel). Though, if the ADC has only one physical channel, the
data that would come from the second ADC channel comes in as noise and has
to be discarded. Because of that, only half of data collected by DMA blocks
shall be passed to IIO buffers when using single-channel ADCs such as
ADAQ4216.

Because only half of the data pushed to DMA is used, the filtered dmaengine
buffer variant allocates blocks that are two times the usual size so the
IIO buffer can still be filled with the data from one DMA block. Otherwise,
a data glitch occurs right in the middle of the IIO buffer.

Add a separate dmaengine buffer variant for filtering data from ADAQ4216
and similar single-channel ADCs supported by the ad4630.c driver.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>