spi: spi-axi-spi-engine: set dtr_caps when DTR is supported
SPI-Engine can be synthesized to support Double Transfer Rate (DTR), in
which case the controller will be capable of sampling data both on the
rising and falling edges of the clock signal. DTR is sometimes advertised
as Dual or Double Data Rate (DDR) and is used to specify when two data bit
transitions happen per clock cycle. In other words, the data in the MISO
line is valid both on the rising and falling edges of SCLK. In DTR mode,
transfers use half the number of SCLK pulses to clock data in comparison to
the Single Data Rate (SDR) mode typically ran by many SPI controllers.

When applicable, fill struct spi_controller dtr_caps field to indicate
SPI-Engine supports DTR.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>